+++ packages/core/testing/src/fake_async.ts
@@ -5,10 +5,12 @@
  * Use of this source code is governed by an MIT-style license that can be
  * found in the LICENSE file at https://angular.io/license
  */
-import {discardPeriodicTasksFallback, fakeAsyncFallback, flushFallback, flushMicrotasksFallback, resetFakeAsyncZoneFallback, tickFallback} from './fake_async_fallback';
-
 const _Zone: any = typeof Zone !== 'undefined' ? Zone : null;
 const fakeAsyncTestModule = _Zone && _Zone[_Zone.__symbol__('fakeAsyncTest')];
+
+const fakeAsyncTestModuleNotLoadedErrorMessage =
+    `zone-testing.js is needed for the async() test helper but could not be found.
+        Please make sure that your environment includes zone.js/dist/zone-testing.js`;
 
 /**
  * Clears out the shared fake async zone for a test.
@@ -19,9 +21,8 @@ const fakeAsyncTestModule = _Zone && _Zone[_Zone.__symbol__('fakeAsyncTest')];
 export function resetFakeAsyncZone(): void {
   if (fakeAsyncTestModule) {
     return fakeAsyncTestModule.resetFakeAsyncZone();
-  } else {
-    return resetFakeAsyncZoneFallback();
   }
+  throw new Error(fakeAsyncTestModuleNotLoadedErrorMessage);
 }
 
 /**
@@ -46,9 +47,8 @@ export function resetFakeAsyncZone(): void {
 export function fakeAsync(fn: Function): (...args: any[]) => any {
   if (fakeAsyncTestModule) {
     return fakeAsyncTestModule.fakeAsync(fn);
-  } else {
-    return fakeAsyncFallback(fn);
   }
+  throw new Error(fakeAsyncTestModuleNotLoadedErrorMessage);
 }
 
 /**
@@ -108,9 +108,8 @@ export function tick(
     }): void {
   if (fakeAsyncTestModule) {
     return fakeAsyncTestModule.tick(millis, tickOptions);
-  } else {
-    return tickFallback(millis, tickOptions);
   }
+  throw new Error(fakeAsyncTestModuleNotLoadedErrorMessage);
 }
 
 /**
@@ -126,9 +125,8 @@ export function tick(
 export function flush(maxTurns?: number): number {
   if (fakeAsyncTestModule) {
     return fakeAsyncTestModule.flush(maxTurns);
-  } else {
-    return flushFallback(maxTurns);
   }
+  throw new Error(fakeAsyncTestModuleNotLoadedErrorMessage);
 }
 
 /**
@@ -139,9 +137,8 @@ export function flush(maxTurns?: number): number {
 export function discardPeriodicTasks(): void {
   if (fakeAsyncTestModule) {
     return fakeAsyncTestModule.discardPeriodicTasks();
-  } else {
-    discardPeriodicTasksFallback();
   }
+  throw new Error(fakeAsyncTestModuleNotLoadedErrorMessage);
 }
 
 /**
@@ -152,7 +149,6 @@ export function discardPeriodicTasks(): void {
 export function flushMicrotasks(): void {
   if (fakeAsyncTestModule) {
     return fakeAsyncTestModule.flushMicrotasks();
-  } else {
-    return flushMicrotasksFallback();
   }
+  throw new Error(fakeAsyncTestModuleNotLoadedErrorMessage);
 }
